Final Project/requirements.txt
python-dotenv==1.0.0
requests==2.31.0
feedparser==6.0.12
openai>=1.0.0,<2.0.0
httpx==0.24.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.100.0
schedule==1.2.0
